Refactored cryptography classes
